requirements.txt
asgiref>=3.2.7
asn1crypto>=0.24.0
attrs>=17.4.0
blinker>=1.4
certifi>=2020.6.20
chardet>=3.0.4
configobj>=5.0.6
constantly>=15.1.0
config>=0.5.0
Deprecated>=1.2.10
gitdb>=4.0.5
GitPython>=3.1.8
hyperlink>=17.3.1
idna>=2.10
incremental>=16.10.1
Interactive>=1.0.1
jsonpatch>=1.16
jsonpointer>=1.10
jsonschema>=2.6.0
keyring>=10.6.0
keyrings.alt>=3.0
MarkupSafe>=1.0
netifaces>=0.10.4
oauthlib>=2.0.6
paho-mqtt
paramiko>=2.7.1
pubsub>=0.1.2
pyasn1>=0.4.2
pyasn1-modules>=0.2.1
PyJWT>=1.7.1
pyOpenSSL>=17.5.0
Pypubsub>=4.0.3
pyserial>=3.4
python-crontab>=2.5.1
python-dateutil>=2.8.1
pytimeparse>=1.1.8
pytz>=2020.1
requests>=2.24.0
requests-unixsocket>=0.1.5
RPi.GPIO>=0.7.0
SecretStorage>=2.3.1
service-identity>=16.0.0
simplejson==3.17.2
